fix ci

+++ src/vectorstore.cpp
@@ -541,7 +541,7 @@ namespace utils
     std::string index2words(int value)
     {
         const static std::vector<std::string> names = {"first", "second", "third", "fourth", "fifth", "sixth", "seventh", "eighth", "ninth", "tenth"};
-        CHATLLM_CHECK((0 <= value) && (value < names.size()));
+        CHATLLM_CHECK((0 <= value) && ((size_t)value < names.size()));
         return names[value];
     }
